+++ src/formatTimestamp.ts
@@ -3,7 +3,7 @@ import { FormatOptions } from '.'
 
 export function formatTimestamp(
   timestamp: number,
-  options: FormatOptions = { format: 'srt' }
+  options: FormatOptions = { format: 'SRT' }
 ) {
   const date = new Date(0, 0, 0, 0, 0, 0, timestamp)
 
@@ -13,6 +13,6 @@ export function formatTimestamp(
   const ms = timestamp - (hours * 3600000 + minutes * 60000 + seconds * 1000)
 
   return `${padLeft(hours)}:${padLeft(minutes)}:${padLeft(seconds)}${
-    options.format === 'vtt' ? '.' : ','
+    options.format === 'WebVTT' ? '.' : ','
   }${padLeft(ms, 3)}`
 }